callback on wifi timeout

+++ platformio-shared-lib/ClonedLibrary/filewatcher/copylibs.py
@@ -21,25 +21,50 @@ def copy_all():
                 workspaceFolderPath, "platformio.ini")
             libDirectory = os.path.abspath(
                 os.path.join(workspaceFolderPath, "src\\lib"))
+
             if (os.path.exists(platformIniPath) and libsLocation.lower() not in libDirectory.lower()):
-                if (os.path.exists(libDirectory)):
-                    try:
-                        shutil.rmtree(libDirectory)
-                    except:
-                        pass
+                with open(platformIniPath, 'r') as platformIniFile:
+                    platformIniStr = platformIniFile.read()
+                    if ";skiplibs" in platformIniStr or "; skiplibs" in platformIniStr:
+                        print("skiplib in "+os.path.basename(workspaceFolderPath))
+                        continue
+                    if (os.path.exists(libDirectory)):
+                        try:
+                            shutil.rmtree(libDirectory)
+                        except:
+                            pass
 
-                copy_to_workspace(workspaceFolderPath, libDirectory)
+                    copy_to_workspace(workspaceFolderPath,
+                                      libDirectory, platformIniStr)
 
 
-def copy_to_workspace(workspaceFolderPath, libDirectory):
-    os.mkdir(libDirectory)
+def copy_to_workspace(workspaceFolderPath, libDirectory, platformIniStr):
+    try:
+        os.mkdir(libDirectory)
+    except:
+        pass
+
+    whitelisted = ";only " in platformIniStr
+
+    def is_whitelisted(file: str):
+        if (not whitelisted):
+            return True
+        basefile = file.split(".")[0]
+        if (";only "+basefile in platformIniStr):
+            return True
+        return False
 
     for file in os.listdir(libsLocation):
+
+        if (not is_whitelisted(file)):
+            continue
         if not file.endswith("main.cpp") and not file.endswith("creds.h"):
             fileCurrent = os.path.join(libsLocation, file)
             shutil.copy(fileCurrent, os.path.join(
                 libDirectory, file))
     for file in os.listdir(libsLibLocation):
+        if (not is_whitelisted(file)):
+            continue
         if not file.endswith("main.cpp") and not file.endswith("creds.h"):
             fileCurrent = os.path.join(libsLibLocation, file)
             shutil.copy(fileCurrent, os.path.join(
